refactor: move isCategoryValidToCreateExpense to repository
